requirements.txt
streamlit
PyJWT
streamlit_authenticator
extra_streamlit_components
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
requests
boto3
python-dotenv
pydantic
passlib[argon2]